// Frontend/grader-frontend/src/lib/api/generated/models/StudentInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Grader
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudentInfo
 */
export interface StudentInfo {
    /**
     * 
     * @type {string}
     * @memberof StudentInfo
     */
    studentId: string;
    /**
     * 
     * @type {string}
     * @memberof StudentInfo
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof StudentInfo
     */
    section: number;
    /**
     * 
     * @type {string}
     * @memberof StudentInfo
     */
    group: string;
    /**
     * url
     * @type {string}
     * @memberof StudentInfo
     */
    picture: string;
    /**
     * 
     * @type {number}
     * @memberof StudentInfo
     */
    score: number;
    /**
     * 
     * @type {number}
     * @memberof StudentInfo
     */
    maxScore: number;
    /**
     * 
     * @type {boolean}
     * @memberof StudentInfo
     */
    withdrawal: boolean;
}

/**
 * Check if a given object implements the StudentInfo interface.
 */
export function instanceOfStudentInfo(value: object): value is StudentInfo {
    if (!('studentId' in value) || value['studentId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('section' in value) || value['section'] === undefined) return false;
    if (!('group' in value) || value['group'] === undefined) return false;
    if (!('picture' in value) || value['picture'] === undefined) return false;
    if (!('score' in value) || value['score'] === undefined) return false;
    if (!('maxScore' in value) || value['maxScore'] === undefined) return false;
    if (!('withdrawal' in value) || value['withdrawal'] === undefined) return false;
    return true;
}

export function StudentInfoFromJSON(json: any): StudentInfo {
    return StudentInfoFromJSONTyped(json, false);
}

export function StudentInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'studentId': json['student_id'],
        'name': json['name'],
        'section': json['section'],
        'group': json['group'],
        'picture': json['picture'],
        'score': json['score'],
        'maxScore': json['max_score'],
        'withdrawal': json['withdrawal'],
    };
}

export function StudentInfoToJSON(json: any): StudentInfo {
    return StudentInfoToJSONTyped(json, false);
}

export function StudentInfoToJSONTyped(value?: StudentInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'student_id': value['studentId'],
        'name': value['name'],
        'section': value['section'],
        'group': value['group'],
        'picture': value['picture'],
        'score': value['score'],
        'max_score': value['maxScore'],
        'withdrawal': value['withdrawal'],
    };
}